refactor(coolify): one copy of how a client and a connection are read
getClient and readConnectionState existed twice, once for the handlers and
once for the deploy pipeline, differing only in whether a cancellation
signal was passed. Two copies of "what an absent row means" is two answers
waiting to disagree about what the user is connected to.

Also: saveToken now spreads the stored settings as clearToken already did,
so a field added to CoolifySchema later is not dropped by whichever runs;
the deploy-key list from GitHub is parsed inside a guard, since a proxy
page answers 200 with HTML and left bare that reaches the user as a
SyntaxError naming neither GitHub nor the key; the db.update mock went,
which nothing could reach; and the insecure-address warning moved out of
components, having no React in it.
